requirements.txt
django>=5.0
 djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-decouple
django-filter
dj-database-url 